+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] / 2026-01-19
+## [1.1.0] / 2026-01-19 - 2025-01-20
 ### Features
 - Support ignore isolator for specific class using attribute `[Isolator(" ")]`.
 - Support copy properties in `BaseType` before and after execute methods. (Fix: #4)